requirements.txt
openai==1.3.6
pandas>=2.0.0
numpy>=1.24.0
konlpy==0.6.0
faiss-cpu>=1.7.0
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
openpyxl>=3.0.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0
httpx==0.24.1